Separate styles and moved utils
