+++ src/App.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+
+import './App.scss';
 import { TodoApp } from './components/TodoApp/TodoApp';
 import { TodosContext } from './components/TodosContext';
