+++ src/components/Header/Header.scss
@@ -1,0 +1,10 @@
+@import "../../utils/mixins.scss";
+
+.new-todo {
+  @include editTitle;
+
+  padding: 16px 16px 16px 60px;
+  border: none;
+  background: rgba(0, 0, 0, 0.01);
+  box-shadow: inset 0 -2px 1px rgba(0,0,0,0.03);
+}
