+++ src/components/TodoItem/TodoItem.scss
@@ -1,0 +1,5 @@
+@import "../../utils/mixins.scss";
+
+.edit {
+  @include editTitle;
+}
