+++ src/components/TodoList/TodoList.tsx
@@ -1,4 +1,6 @@
 import React from 'react';
+
+import './TodoList.scss';
 import { Todo } from '../../types/Todo';
 import { TodoItem } from '../TodoItem';
