+++ src/index.tsx
@@ -1,9 +1,5 @@
 import { createRoot } from 'react-dom/client';
 
-import './styles/index.css';
-import './styles/todo-list.css';
-import './styles/filters.css';
-
 import { App } from './App';
 
 const container = document.getElementById('root') as HTMLDivElement;
